update storage connector
